network-libp2p: Remove unused code
Remove some unused code that was located into the codecs directory
and move the `typed` codec to be the only code into the `codecs`
submodule.
