--- network-libp2p/src/dispatch/codecs/tokio_adapter.rs
@@ -1,118 +1,0 @@
-// TODO: Move to nimiq-utils?
-
-use std::{
-    io,
-    pin::Pin,
-    task::{Context, Poll},
-};
-
-use futures::ready;
-use pin_project_lite::pin_project;
-
-pin_project! {
-    #[derive(Clone, Debug)]
-    pub struct TokioAdapter<T> {
-        #[pin]
-        inner: T,
-    }
-}
-
-impl<T> TokioAdapter<T> {
-    pub fn new(inner: T) -> Self {
-        Self { inner }
-    }
-
-    pub fn into_inner(self) -> T {
-        self.inner
-    }
-
-    pub fn get_ref(&self) -> &T {
-        &self.inner
-    }
-
-    pub fn get_mut(&mut self) -> &mut T {
-        &mut self.inner
-    }
-}
-
-impl<T> futures::AsyncRead for TokioAdapter<T>
-where
-    T: tokio::io::AsyncRead,
-{
-    fn poll_read(
-        self: Pin<&mut Self>,
-        cx: &mut Context<'_>,
-        buf: &mut [u8],
-    ) -> Poll<io::Result<usize>> {
-        let mut buf = tokio::io::ReadBuf::new(buf);
-        ready!(tokio::io::AsyncRead::poll_read(
-            self.project().inner,
-            cx,
-            &mut buf
-        ))?;
-        Poll::Ready(Ok(buf.filled().len()))
-    }
-}
-
-impl<T> futures::AsyncWrite for TokioAdapter<T>
-where
-    T: tokio::io::AsyncWrite,
-{
-    fn poll_write(
-        self: Pin<&mut Self>,
-        cx: &mut Context<'_>,
-        buf: &[u8],
-    ) -> Poll<std::io::Result<usize>> {
-        tokio::io::AsyncWrite::poll_write(self.project().inner, cx, buf)
-    }
-
-    fn poll_flush(self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<io::Result<()>> {
-        tokio::io::AsyncWrite::poll_flush(self.project().inner, cx)
-    }
-
-    fn poll_close(self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<io::Result<()>> {
-        tokio::io::AsyncWrite::poll_shutdown(self.project().inner, cx)
-    }
-}
-
-impl<T> tokio::io::AsyncRead for TokioAdapter<T>
-where
-    T: futures::AsyncRead,
-{
-    fn poll_read(
-        self: Pin<&mut Self>,
-        cx: &mut Context<'_>,
-        buf: &mut tokio::io::ReadBuf<'_>,
-    ) -> Poll<io::Result<()>> {
-        // TODO: It's not optimal to initialize the buffer
-        let buffer = buf.initialize_unfilled();
-        let n = ready!(futures::AsyncRead::poll_read(
-            self.project().inner,
-            cx,
-            buffer
-        ))?;
-        buf.advance(n);
-        Poll::Ready(Ok(()))
-    }
-}
-
-impl<T> tokio::io::AsyncWrite for TokioAdapter<T>
-where
-    T: futures::AsyncWrite,
-{
-    fn poll_write(
-        self: Pin<&mut Self>,
-        cx: &mut Context<'_>,
-        buf: &[u8],
-    ) -> Poll<io::Result<usize>> {
-        futures::AsyncWrite::poll_write(self.project().inner, cx, buf)
-    }
-
-    fn poll_flush(self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<io::Result<()>> {
-        futures::AsyncWrite::poll_flush(self.project().inner, cx)
-    }
-
-    fn poll_shutdown(self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<io::Result<()>> {
-        futures::AsyncWrite::poll_close(self.project().inner, cx)
-    }
-}
